requirements.txt
# python 3.10 environment
numpy<2.0
torch==2.2.2
torch-geometric==2.6.1
torchvision==0.17.2
pybind11==2.13.6
pybind11-global==2.13.6
pyyaml==6.0.2